--- src/main/java/com/acme/edu/DataLoader.java
@@ -1,9 +1,0 @@
-package com.acme.edu;
-
-/**
- *
- * @author aleksandr-nevsky
- */
-public class DataLoader {
-
-}
