--- src/main/java/su/jet/java/edu/loader/impl/XmlDataReader.java
@@ -1,17 +1,0 @@
-package su.jet.java.edu.loader.impl;
-
-import su.jet.java.edu.loader.interfaces.DataReader;
-import su.jet.java.edu.loader.UserData;
-
-/**
- *
- * @author aleksandr-nevsky
- */
-public class XmlDataReader implements DataReader {
-
-    @Override
-    public UserData read() {
-        return new UserData(2, "Vasya xml");
-    }
-
-}
